// src/types/beats/ripple.ts
export interface RippleLedgerTransactionsMetadata {
  AffectedNodes: RippleLedgerTransactionsMetadataAffectednodes[];
  TransactionIndex: number;
  TransactionResult: string;
  delivered_amount?: Record<string, unknown> | string;
  DeliveredAmount?: Record<string, unknown> | string;
}

export interface RippleLedgerTransactionsMemosMemo {
  MemoData: string;
  MemoType?: string;
  MemoFormat?: string;
}

export interface RippleLedgerTransactionsMemos {
  Memo: RippleLedgerTransactionsMemosMemo;
}

export interface RippleLedgerTransactionsPaths {
  account: string;
  type: number;
  currency: string;
  issuer: string;
}

export interface RippleLedgerTransactionsLimitamount {
  currency: string;
  issuer: string;
  value: string;
}

export interface RippleLedgerTransactionsSignersSigner {
  Account: string;
  SigningPubKey: string;
  TxnSignature: string;
}

export interface RippleLedgerTransactionsSigners {
  Signer?: RippleLedgerTransactionsSignersSigner;
}

export interface RippleLedgerTransactionsAsset {
  currency: string;
}

export interface RippleLedgerTransactionsAsset2 {
  currency: string;
  issuer: string;
}

export interface RippleLedgerTransactionsLptokenin {
  currency: string;
  issuer: string;
  value: string;
}

export interface RippleLedgerTransactions {
  Account: string;
  Fee: string;
  Flags: number;
  LastLedgerSequence: number;
  Sequence: number;
  SigningPubKey: string;
  TakerGets: Record<string, unknown> | string;
  TakerPays: Record<string, unknown> | string;
  TransactionType: string;
  TxnSignature: string;
  hash: string;
  metaData: RippleLedgerTransactionsMetadata;
  Amount: Record<string, unknown> | string;
  Destination: string;
  DestinationTag: number;
  DeliverMax: Record<string, unknown> | string;
  NFTokenID: string;
  TicketSequence: number;
  SourceTag: number;
  OfferSequence: number;
  DeliverMin: Record<string, unknown> | string;
  SendMax: Record<string, unknown> | string;
  TicketCount: number;
  Memos: RippleLedgerTransactionsMemos[];
  Expiration: number;
  InvoiceID?: string;
  Paths: RippleLedgerTransactionsPaths[][];
  LimitAmount: RippleLedgerTransactionsLimitamount;
  Signers?: RippleLedgerTransactionsSigners[];
  Asset?: RippleLedgerTransactionsAsset;
  Asset2?: RippleLedgerTransactionsAsset2;
  LPTokenIn?: RippleLedgerTransactionsLptokenin;
  Owner?: string;
}

export interface RippleLedgerDiff {
  object_id: string;
  object: Record<string, unknown> | string;
}

export interface RippleLedger {
  account_hash: string;
  close_flags: number;
  close_time: number;
  close_time_human: string;
  close_time_resolution: number;
  close_time_iso: string;
  ledger_hash: string;
  parent_close_time: number;
  parent_hash: string;
  total_coins: string;
  transaction_hash: string;
  ledger_index: string;
  closed: boolean;
  transactions: RippleLedgerTransactions[];
  diff: RippleLedgerDiff[];
  _network: string;
  _vm: string;
}